+++ src/main/java/com/emsi/salesmasterbe2/services/impl/FactureServiceImpl.java
@@ -2,15 +2,19 @@ package com.emsi.salesmasterbe2.services.impl;
 
 import com.emsi.salesmasterbe2.daos.FactureDao;
 import com.emsi.salesmasterbe2.entities.Facture;
+import com.emsi.salesmasterbe2.payload.response.PagedResponse;
 import com.emsi.salesmasterbe2.repository.FactureRepository;
 import com.emsi.salesmasterbe2.services.FactureService;
+import com.emsi.salesmasterbe2.utils.AppConstants;
+import com.emsi.salesmasterbe2.utils.AppUtils;
 import com.emsi.salesmasterbe2.utils.ObjectMapperUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class FactureServiceImpl implements FactureService {
@@ -24,27 +28,42 @@ public class FactureServiceImpl implements FactureService {
 
     @Override
     public FactureDao saveFacture(FactureDao factureDao) {
-        Facture factureEntity = ObjectMapperUtils.map(factureDao,Facture.class);
+        Facture factureEntity = ObjectMapperUtils.map(factureDao, Facture.class);
         factureEntity = factureRepository.save(factureEntity);
-        return ObjectMapperUtils.map(factureEntity,FactureDao.class);
+        return ObjectMapperUtils.map(factureEntity, FactureDao.class);
     }
 
     @Override
     public FactureDao getFactureById(Long id) {
         Optional<Facture> factureOptional = factureRepository.findById(id);
-        return ObjectMapperUtils.map(factureOptional.get(),FactureDao.class);
+        if (factureOptional.isEmpty()) {
+            throw new IllegalArgumentException("Facture with ID " + id + " not found");
+        }
+        return ObjectMapperUtils.map(factureOptional.get(), FactureDao.class);
     }
 
     @Override
-    public List<FactureDao> getAllFactures() {
-        List<Facture> factures = factureRepository.findAll();
-        return ObjectMapperUtils.mapAll(factures,FactureDao.class);
+    public PagedResponse<FactureDao> getAllFactures(int page, int size) {
+        AppUtils.validatePageNumberAndSize(page, size);
+        Pageable pageable = PageRequest.of(page, size);
+        Page<Facture> facturesPage = factureRepository.findAll(pageable);
+        return new PagedResponse<>(
+                ObjectMapperUtils.mapAll(facturesPage.getContent(), FactureDao.class),
+                page,
+                size,
+                facturesPage.getNumberOfElements(),
+                facturesPage.getTotalPages()
+        );
     }
 
     @Override
-    public void deleteFacture(Long id) {
-        factureRepository.deleteById(id);
+    public FactureDao deleteFacture(Long id) {
+        Optional<Facture> factureOptional = factureRepository.findById(id);
+        if (factureOptional.isPresent()) {
+            factureRepository.deleteById(id);
+            return ObjectMapperUtils.map(factureOptional.get(), FactureDao.class);
+        } else {
+            throw new IllegalArgumentException("Facture with ID " + id + " not found");
+        }
     }
-
-
 }
