+++ src/main/java/com/emsi/salesmasterbe2/services/impl/ProduitServiceImpl.java
@@ -2,15 +2,18 @@ package com.emsi.salesmasterbe2.services.impl;
 
 import com.emsi.salesmasterbe2.daos.ProduitDao;
 import com.emsi.salesmasterbe2.entities.Produit;
+import com.emsi.salesmasterbe2.payload.response.PagedResponse;
 import com.emsi.salesmasterbe2.repository.ProduitRepository;
 import com.emsi.salesmasterbe2.services.ProduitService;
+import com.emsi.salesmasterbe2.utils.AppUtils;
 import com.emsi.salesmasterbe2.utils.ObjectMapperUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class ProduitServiceImpl implements ProduitService {
@@ -32,18 +35,34 @@ public class ProduitServiceImpl implements ProduitService {
     @Override
     public ProduitDao getProduitById(Long id) {
         Optional<Produit> produitOptional = produitRepository.findById(id);
+        if (produitOptional.isEmpty()) {
+            throw new IllegalArgumentException("Produit with ID " + id + " not found");
+        }
         return ObjectMapperUtils.map(produitOptional.get(), ProduitDao.class);
     }
 
     @Override
-    public List<ProduitDao> getAllProduits() {
-        List<Produit> produits = produitRepository.findAll();
-        return ObjectMapperUtils.mapAll(produits,ProduitDao.class);
+    public PagedResponse<ProduitDao> getAllProduits(int page, int size) {
+        AppUtils.validatePageNumberAndSize(page, size);
+        Pageable pageable = PageRequest.of(page, size);
+        Page<Produit> produitsPage = produitRepository.findAll(pageable);
+        return new PagedResponse<>(
+                ObjectMapperUtils.mapAll(produitsPage.getContent(), ProduitDao.class),
+                page,
+                size,
+                produitsPage.getNumberOfElements(),
+                produitsPage.getTotalPages()
+        );
     }
 
     @Override
-    public void deleteProduit(Long id) {
-        produitRepository.deleteById(id);
+    public ProduitDao deleteProduit(Long id) {
+        Optional<Produit> produitOptional = produitRepository.findById(id);
+        if (produitOptional.isPresent()) {
+            produitRepository.deleteById(id);
+            return ObjectMapperUtils.map(produitOptional.get(), ProduitDao.class);
+        } else {
+            throw new IllegalArgumentException("Produit with ID " + id + " not found");
+        }
     }
-
 }
